6ème commit

+++ src/main/java/fr/algorithmie/CalculMoyenne.java
@@ -4,9 +4,9 @@ public class CalculMoyenne {
 
 	public static void main(String[] args) {
 		int[] array = {1, 15, -3, 0, 8, 7, 4, -2, 28, 7, -1, 17, 2, 3, 0, 14, -4} ;
-		double moyenneValeurs = 0.0;
+		float moyenneValeurs = 0.0f;
 		for (int i = 0; i < array.length; i++) {
-			moyenneValeurs = moyenneValeurs + (double)array [i] /array.length ;
+			moyenneValeurs = moyenneValeurs + (float)array [i] /array.length ;
 		}
 		System.out.println("La moyenne est : " + moyenneValeurs);
